Fixed some errors with spawning enemies and follow behavior
Also hid the black bars, temporarily.  Several scripts' public objects linked to specific "in-scene" instances of objects instead of their prefab equivalents, which is why we were having dogs stop spawning once one died.  Also, the player object that the dogs follow is now a private member variable in EnemyDog, passed down from a public member in Waves.  This way, the enemy will always be inhereting the proper instance of the Player object.